requirements.txt
asgiref==3.7.2
black==23.11.0
boto3==1.29.7
botocore==1.32.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==21.0.0
html2text==2020.1.16
idna==3.6
inflection==0.5.1
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.8.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.0
requests~=2.31.0
xmltodict~=0.13.0
